ADDED: corfu.el - Completion Overlay Region FUnction.
Corfu enhances completion at point with a small completion popup. The current candidates are shown in a popup below or above the point. Corfu is the minimalistic completion-in-region counterpart of the Vertico minibuffer UI.

Corfu is a small package, which relies on the Emacs completion facilities and concentrates on providing a polished completion UI. Completions are either provided by commands like dabbrev-completion or by pluggable backends (completion-at-point-functions, Capfs). Most programming language major modes implement a Capf. Furthermore the language server packages, Eglot and Lsp-mode, use Capfs which talk to the LSP server to retrieve the completions. Corfu does not include its own completion backends. The Emacs built-in Capfs and the Capfs provided by other programming language packages are usually sufficient. A few additional Capfs and completion utilities are provided by the Cape package.

NOTE: Corfu uses child frames to show the popup and falls back to the
default setting of the completion-in-region-function on non-graphical
displays. If you want to use Corfu in the terminal, install the
package corfu-terminal, which provides an alternative overlay-based
display.
https://github.com/minad/corfu.git