ADDED: cape.el - Completion At Point Extensions.
MELPA Stable
Cape provides Completion At Point Extensions which can be used in combination with the Corfu completion UI or the default completion UI. The completion backends used by completion-at-point are so called completion-at-point-functions (Capfs). In principle, the Capfs provided by Cape can also be used by Company.

You can register the cape-* functions in the completion-at-point-functions list. This makes the backends available for completion, which is usually invoked by pressing TAB or M-TAB. The functions can also be invoked interactively to trigger the respective completion at point. You can bind them directly to a key in your user configuration. Notable commands/Capfs are cape-line for completion of a line from the current buffer and cape-file for completion of a file name. The command cape-symbol is particularly useful for documentation of Elisp packages or configurations, since it completes Elisp symbols anywhere.

Cape has the super power to transform Company backends into Capfs and
merge multiple Capfs into a Super-Capf! These transformers allow you
to still take advantage of Company backends even if you are not using
Company as frontend.
https://github.com/minad/cape.git