ADDED: benchmark-init.el benchmark your Emacs initialization.
This is a simple benchmark of calls to Emacs require and load functions. It can be used to keep track of where time is being spent during Emacs startup in order to optimize startup times.

The code was originally based on init-benchmarking.el by Steve Purcell with many additional modifications having gone into it since.

https://github.com/dholm/benchmark-init-el.git