ADDED: TempEl - Simple templates for Emacs.
Tempel is a tiny template package for Emacs, which uses the syntax of
the Emacs Tempo library. Tempo is an ancient temple of the church of
Emacs. It is 27 years old, but still in good shape since it
successfully resisted change over the decades. However it may look a
bit dusty here and there. Therefore we present Tempel, a modernized
implementation of Tempo. Tempel integrates well with the standard
completion-at-point-functions mechanism which is used by Emacs for
in-buffer completion.

https://github.com/minad/tempel.git